Add changeable word count and guess count
Word count can now be changed via increment and decrement buttons. Guess count updates itself accordingly.

Word count is constrained to values from 1 to 10. Guess count is calculated as a function of the word count, so it automatically updates as the word count changes. The displayed word grids update accordingly.